+++ src/interfaces/IOrbsToken.sol
@@ -6,9 +6,9 @@ import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 interface IOrbsToken is IERC20 {
     function mint(uint256 amount) external;
 
-    function startDripping(address addr, uint256 multiplier) external;
+    function startDripping(address addr, uint128 multiplier) external;
 
-    function stopDripping(address addr, uint256 multiplier) external;
+    function stopDripping(address addr, uint128 multiplier) external;
 
     function burn(address from, uint256 value) external;
 }
